requirements.txt
django==3.2
djangorestframework==3.12.4
psycopg2==2.8.6
python-decouple==3.4
django-cors-headers==3.7.0
dj-rest-auth==2.1.5
djangorestframework-simplejwt==4.7.0